Add more information on active team roster
Add the data symbol TEAM_ACTIVE_ROSTER. This is the list of all active
members on the player's team. This subsumes a couple data symbols
related to the leader/party members.

Also fill in a bunch of fields on struct team_member, and do some
necessary reorganization. This includes moving `struct item` and `struct
move` into dungeon_mode_common.h so that they're accessible from
common.h.
