+++ headers/data/ram.h
@@ -27,9 +27,7 @@ extern uint32_t PLAY_TIME_SECONDS;
 extern uint8_t PLAY_TIME_FRAME_COUNTER;
 extern char TEAM_NAME[10];
 extern struct ground_monster TEAM_MEMBER_LIST[555];
-extern uint32_t LEADER_IQ_SKILLS[3];
-extern char LEADER_NICKNAME[10];
-extern uint32_t PARTY_MEMBER_2_IQ_SKILLS[3];
+extern struct team_member TEAM_ACTIVE_ROSTER[7];
 extern uint32_t FRAMES_SINCE_LAUNCH_TIMES_THREE;
 extern bool TURNING_ON_THE_SPOT_FLAG;
 extern struct floor_generation_status FLOOR_GENERATION_STATUS;
