+++ headers/types/dungeon_mode/dungeon_mode.h
@@ -4,6 +4,7 @@
 #define HEADERS_TYPES_DUNGEON_MODE_H_
 
 #include "enums.h"
+#include "dungeon_mode_common.h"
 
 // Used in various contexts, like with entity positions in the dungeon
 struct position {
@@ -11,25 +12,6 @@ struct position {
     int16_t y;
 };
 ASSERT_SIZE(struct position, 4);
-
-// Item info
-struct item {
-    // 0x0: flags: 1-byte bitfield
-    bool f_exists : 1;  // Validity flag
-    bool f_in_shop : 1; // In a Kecleon Shop
-    bool f_unpaid : 1;  // Picked up from a Kecleon Shop but not paid for yet
-    bool f_sticky : 1;  // Sticky
-    bool f_set : 1;     // Usable by L+R
-    uint8_t flags_unk5 : 3;
-
-    // 0x1: For bag items. 0 for none, 1 if held by the leader, 2 for the second party member, etc.
-    uint8_t held_by;
-    // 0x2: Only for stackable items. Will be 0 if unapplicable. For Poké, this is an "amount code"
-    // rather than the literal amount
-    uint16_t quantity;
-    struct item_id_16 id; // 0x4
-};
-ASSERT_SIZE(struct item, 6);
 
 // Trap info
 struct trap {
@@ -46,38 +28,6 @@ struct trap {
     undefined field_0x3;
 };
 ASSERT_SIZE(struct trap, 4);
-
-// Monster move info
-struct move {
-    // 0x0: flags0: 1-byte bitfield
-    bool f_exists : 1;                   // This move will show up in the move list
-    bool f_subsequent_in_link_chain : 1; // This move is in a link chain, but not the first move
-    bool f_enabled_for_ai : 1;           // AI can use this move
-    bool f_set : 1;                      // Usable by L+A
-    bool f_last_used : 1; // This move was the last one used (for things like Torment?)
-    bool f_disabled : 1;  // For Torment (and maybe others)
-    uint8_t flags0_unk6 : 2;
-
-    undefined field_0x1;
-
-    // 0x2: flags2: 2-byte bitfield
-    bool f_sealed : 1; // Sealed by a Seal Trap. Also prevents AI from using this move
-    uint16_t flags2_unk1 : 2;
-    bool f_consume_pp : 1;   // This move will consume PP this turn
-    bool f_consume_2_pp : 1; // Consume 2 PP this turn. No effect unless f_consume_pp is set
-    uint16_t flags2_unk5 : 1;
-    // Consume 4 PP this turn. No effect unless f_consume_pp is set, overrides f_consume_2_pp
-    bool f_consume_4_pp : 1;
-    uint16_t flags2_unk7 : 1;
-    uint16_t flags2_unk8 : 1;
-    bool f_exclusive_item_pp_boost : 1; // A PP-boosting exclusive item is in effect
-    uint16_t flags3_unk10 : 6;
-
-    struct move_id_16 id; // 0x4
-    uint8_t pp;           // 0x6: Current PP
-    uint8_t ginseng;      // 0x7: Ginseng boost
-};
-ASSERT_SIZE(struct move, 8);
 
 // Used to hold data during damage calculation
 struct damage_data {
@@ -358,12 +308,12 @@ struct monster {
     undefined field_0x5b;
     undefined field_0x5c;
     undefined field_0x5d;
-    struct type_id_8 type1;       // 0x5E
-    struct type_id_8 type2;       // 0x5F
-    struct ability_id_8 ability1; // 0x60
-    struct ability_id_8 ability2; // 0x61
-    struct item held_item;        // 0x62
-    uint16_t held_item_id;        // 0x68: Appears to be a mirror of held_item.id
+    struct type_id_8 type1;         // 0x5E
+    struct type_id_8 type2;         // 0x5F
+    struct ability_id_8 ability1;   // 0x60
+    struct ability_id_8 ability2;   // 0x61
+    struct item held_item;          // 0x62
+    struct item_id_16 held_item_id; // 0x68: Appears to be a mirror of held_item.id
     // Previous position data is used by the AI
     struct position prev_pos;  // 0x6A: Position 1 turn ago
     struct position prev_pos2; // 0x6E: Position 2 turns ago
